+++ fleet_adapter_kachaka/config.yaml
