+++ fleet_adapter_kachaka/config_template.yaml
