+++ fleet_adapter_kachaka/fleet_adapter_kachaka/delivery_node.py
