+++ fleet_adapter_kachaka/launch/fleet_adapter.launch.xml
@@ -7,12 +7,12 @@
   <arg name="output" default="screen"/>
   <arg name="robot_name" default="kachaka"/>
   <!-- Fleet adapter -->
-  <node pkg="fleet_adapter_for_kachaka" exec="fleet_adapter" args="--config_file $(var config_file) --nav_graph $(var nav_graph_file)" output="both">
+  <node pkg="fleet_adapter_kachaka" exec="fleet_adapter" args="--config_file $(var config_file) --nav_graph $(var nav_graph_file)" output="both">
     <param name="server_uri" value="$(var server_uri)"/>
   </node>
 
   <!-- delivery_node -->
-  <node pkg="fleet_adapter_for_kachaka" exec="delivery_node" args="-c $(var config_file)" output="both">
+  <node pkg="fleet_adapter_kachaka" exec="delivery_node" args="-c $(var config_file)" output="both">
     <param name="robot_name" value="$(var robot_name)"/>
   </node>
