+++ fleet_adapter_kachaka/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>fleet_adapter_for_kachaka</name>
+  <name>fleet_adapter_kachaka</name>
   <version>0.0.0</version>
   <description>An RMF fleet adapter for kachaka</description>
   <maintainer email="yadunund@openrobotics.org">Yadunund</maintainer>
