+++ fleet_adapter_kachaka/setup.cfg
@@ -1,7 +1,7 @@
 [develop]
-script-dir=$base/lib/fleet_adapter_for_kachaka
+script-dir=$base/lib/fleet_adapter_kachaka
 [install]
-install-scripts=$base/lib/fleet_adapter_for_kachaka
+install-scripts=$base/lib/fleet_adapter_kachaka
 [flake8]
 aggressive = 2
 ignore = W503, ANN002, ANN003, ANN1, C801
