+++ fleet_adapter_kachaka/test/test_copyright.py
@@ -12,12 +12,12 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from ament_pep257.main import main
 import pytest
+from ament_copyright.main import main
 
 
+@pytest.mark.copyright
 @pytest.mark.linter
-@pytest.mark.pep257
-def test_pep257():
+def test_copyright():
     rc = main(argv=['.', 'test'])
-    assert rc == 0, 'Found code style errors / warnings'
+    assert rc == 0, 'Found errors'
